# backend-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.TagCondition:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  model.TagConditions:
    properties:
      tags:
        items:
          $ref: '#/definitions/model.TagCondition'
        type: array
    type: object
  model.TagRouting:
    properties:
      bucket:
        type: string
      conditions:
        $ref: '#/definitions/model.TagConditions'
      createdAt:
        type: string
      description:
        type: string
      driver:
        type: string
      extraConfig:
        type: string
      id:
        type: integer
      isDefault:
        type: boolean
      matchType:
        type: string
      pathPrefix:
        type: string
      priority:
        type: integer
      ruleName:
        type: string
      status:
        type: integer
      updatedAt:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 后端管理服务 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Server API
  version: "1.0"
paths:
...
  /system/routing-rules:
    get:
      description: 获取所有标签路由规则
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.TagRouting'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取路由规则列表
      tags:
      - 标签路由
    post:
      consumes:
      - application/json
      description: 创建新的标签路由规则
      parameters:
      - description: 路由规则
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.TagRouting'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.TagRouting'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建路由规则
      tags:
      - 标签路由
  /system/routing-rules/{id}:
    delete:
      description: 删除指定标签路由规则
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除路由规则
      tags:
      - 标签路由
    get:
      description: 根据 ID 获取标签路由规则详情
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.TagRouting'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取路由规则详情
      tags:
      - 标签路由
    put:
      consumes:
      - application/json
      description: 更新指定标签路由规则
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 路由规则
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.TagRouting'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.TagRouting'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新路由规则
      tags:
      - 标签路由
  /system/routing-rules/{id}/priority:
    put:
      consumes:
      - application/json
      description: 更新指定标签路由规则优先级
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 优先级
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新路由规则优先级
      tags:
      - 标签路由
  /system/routing-rules/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新指定标签路由规则状态
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新路由规则状态
      tags:
      - 标签路由
  /system/routing-rules/{id}/test:
    post:
      consumes:
      - application/json
      description: 测试指定标签路由规则是否匹配
      parameters:
      - description: 规则 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签列表
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 测试路由规则
      tags:
      - 标签路由
  /system/routing-rules/batch-priority:
    post:
      consumes:
      - application/json
      description: 批量更新标签路由规则优先级
      parameters:
      - description: 优先级映射
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 批量更新路由规则优先级
      tags:
      - 标签路由
  /system/routing-rules/test-route:
    post:
      consumes:
      - application/json
      description: 根据文件信息测试路由结果
      parameters:
      - description: 文件信息
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 测试文件路由
      tags:
      - 标签路由
...
securityDefinitions:
  BearerAuth:
    description: JWT Token（格式：Bearer {token}）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"